+++ Commands/Shortcuts.js
@@ -1,5 +1,3 @@
-const requireDir = require('require-dir');
-const commands = requireDir('.');
 const Logger = require('./Logger');
 
 /**
@@ -7,36 +5,43 @@ const Logger = require('./Logger');
  * @does handels all Shortcuts
  */
 module.exports = { //Redirects Shortcuts to Commands
-    a: function (message) {
-        call('ask', message);
-    },
-    d: function (message) {
-        message.content = 'dorime';
-        call('play', message);
-    },
-    e: function (message) {
-        call('emoji', message);
-    },
-    j: function (message) {
-        call('jap', message);
-    },
-    p: function (message) {
-        call('play', message);
-    },
-    s: function (message) {
-        call('stop', message);
-    },
-    w: function (message) {
-        call('weebiefy', message);
-    }
+    Shortcuts: function (message, commands, shortcut) {
 
-}
+        let command = '';
 
-function call(name, message) {
-    try {
-        commands[name][name](message);
-    } catch (error) {
-        Logger.log(error);
-        message.channel.send('Command not Found, use !help for help');
+        switch (shortcut) {
+            case 'a':
+                command = 'ask';
+                break;
+            case 'd':
+                message.content = 'dorime';
+                command = 'play';
+                break;
+            case 'e':
+                command = 'emoji';
+                break;
+            case 'j':
+                command = 'jap';
+                break;
+            case 'p':
+                command = 'play';
+                break;
+            case 's':
+                command = 'stop';
+                break;
+            case 'w':
+                command = 'weebiefy';
+                break;
+            default:
+                message.channel.send('Command not Found, use !help for help');
+                return;
+        }
+        
+        try {
+            commands[command][command](message);
+        } catch (error) {
+            Logger.log(error);
+            message.channel.send('Command not Found, use !help for help');
+        }
     }
 }