+++ Files/Changelog.md
@@ -1,5 +1,9 @@
 //CHANGE VERSION && SET DEBUG TO FALSE AGAIN
 
+4.1.5: 01.05.2020
+    -Refractored Shortcuts.js
+    -Changed league Detection
+
 4.1.4: 30.04.2020
     -Refractored league Detection
     -Refractored name to botwide Names